requirements.txt
discord.py
yfinance
pandas
numpy
requests
pytz
python-dotenv